Changed search property in defaultGameFilters to empty string'
